Workaround Google Secure LDAP re-bind bug
It seems Google Secure LDAP doesn't properly
honor a re-bind in the same LDAP connection.
The first user is still used for the query despite
a successful second bind.

This breaks Vault LDAP integration as it first
bind as the user to authenticate him, then re-bind
as the given Bind DN to do the group search query.

Since the user doesn't have any permissions to see
the existings groups nor seing his memberOf attribute,
this prevents Groups membership to work.

See https://github.com/hashicorp/vault/issues/6111
for some info on the issue.

The workaround consists in closing/opening the connections
between the two binds as explained in the first message on
the issue.